Merged Versions
I merged your user prompts and my functions. This version is much simpler and lighter. It also only keeps the key in memory and not on disk. It only encrypts and decrypts a file called 'data.txt' right now but eventually it will use a target list. It also still needs to encrypt the symmetric key on disk but it needs the public key from the pair to do that.